backend/pyproject.toml
[tool.poetry]
name = "connectify"
version = "0.1.0"
description = ""
authors = ["DashaVed <dasha.ved93@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"
markdown = "^3.4.3"
django-filter = "^23.1"
django-cors-headers = "^3.14.0"
pillow = "^9.5.0"
djoser = "^2.2.0"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
